requirements.txt
google-cloud-pubsub==2.18.1
google-cloud-vision==3.5.0
google-api-core==1.34.0
proto-plus==1.23.0
apache-beam==2.51.0
google-apitools==0.5.32
google-cloud-bigquery==3.13.0
protobuf==3.20.3
argparse
logging
pandas
haversine
streamlit
folium
streamlit-folium